Listen for login notification and update the store.
