+++ react-native/react/actions/config/index.js
@@ -1,5 +1,4 @@
 import * as Constants from '../../constants/config'
-import {autoLogin} from '../login'
 import engine from '../../engine'
 import * as native from './index.native'
 
@@ -7,22 +6,6 @@ const loggingIncomingMap = {
   'keybase.1.logUi.log': (param, response) => {
     console.log(param)
     response.result()
-  }
-}
-
-export function startup () {
-  return function (dispatch) {
-    // Also call getCurrentStatus if the service goes away/comes back.
-    engine.listenOnConnect('getCurrentStatus', () => dispatch(getCurrentStatus()))
-
-    dispatch({type: Constants.startupLoading})
-
-    dispatch(getCurrentStatus())
-      .then(() => dispatch(getConfig()))
-      .then(() => dispatch({type: Constants.startupLoaded}))
-      .catch(error => {
-        dispatch({type: Constants.startupLoaded, payload: error, error: true})
-      })
   }
 }
 
@@ -54,10 +37,6 @@ export function getCurrentStatus (): (dispatch: Dispatch) => Promise<void> {
           type: Constants.statusLoaded,
           payload: {status}
         })
-
-        if (status.loggedIn) {
-          dispatch(autoLogin())
-        }
 
         resolve()
       })
