+++ react-native/react/nav.android.js
@@ -18,7 +18,7 @@ import {androidTabBarHeight} from './styles/native'
 
 import {switchTab} from './actions/tabbed-router'
 import {navigateBack} from './actions/router'
-import {startup} from './actions/config'
+import {startup} from './actions/startup'
 
 import * as Constants from './constants/config'
